requirements.txt
# ManiSkill环境和仿真依赖
mani-skill>=3.0.0
sapien>=3.0.0

# 强化学习算法库
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数值计算和数据处理
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# 可视化和日志
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0
wandb>=0.12.0

# 系统工具
tqdm>=4.62.0
psutil>=5.8.0
pyyaml>=6.0

# 开发和调试工具
jupyter>=1.0.0
ipython>=7.0.0
pytest>=6.0.0

# 可选：GPU加速 (根据系统选择)
# torch-audio  # 如果需要音频处理
# cupy-cuda11x  # 如果使用CUDA 11.x
# cupy-cuda12x  # 如果使用CUDA 12.x 